rename layer catalog
